+++ src/gemseo_mlearning/active_learning/acquisition_criteria/level_set/ef.py
@@ -18,6 +18,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from numpy import array
+from numpy import atleast_2d
+from numpy import dot
+from numpy import max as np_max
+from numpy import maximum
+from numpy import mean
 from numpy import nan_to_num
 from scipy.stats import norm
 
@@ -28,14 +34,29 @@ from gemseo_mlearning.active_learning.acquisition_criteria.level_set.base_ei_ef 
 if TYPE_CHECKING:
     from gemseo.typing import NumberArray
 
+    from gemseo_mlearning.active_learning.distributions.regressor_distribution import (  # noqa: E501
+        RegressorDistribution,
+    )
+
 
 class EF(BaseEIEF):
     r"""The expected feasibility.
 
-    This acquisition criterion is expressed as:
+    This acquisition criterion is expressed as
 
     $$
-    EF[x] = \mathbb{S}[Y(x)]\times
+    EF[x] =  \mathbb{E}\left[\max(\kappa\mathbb{S}[Y(x)] - |y - Y(x)|,0)\right]
+    $$
+    where $y$ is the model output value characterizing the level set
+    and $Y$ is the random process
+    modelling the uncertainty of the surrogate model $\hat{f}$.
+
+    In the case of a Gaussian regressor,
+    it has an analytic expression:
+
+    $$
+    EF[x] =
+    \mathbb{S}[Y(x)]\times
     (
     \kappa
     (\Phi(t^+)-\Phi(t^-))
@@ -44,16 +65,28 @@ class EF(BaseEIEF):
     )
     $$
 
-    where $Y$ is the random process
-    modelling the uncertainty of the surrogate model $\hat{f}$,
+    where
     $t=\frac{y - \mathbb{E}[Y(x)]}{\mathbb{S}[Y(x)]}$,
     $t^+=t+\kappa$,
     $t^-=t-\kappa$,
     $y$ is the model output value characterizing the level set
-    and $\kappa\in[0,1]$ (default: 2).
+    and $\kappa>0$.
+
+    For the acquisition of $q>1$ points at a time,
+    the acquisition criterion changes to
+    $$EF[x_1,\dots,x_q] = \mathbb{E}\left[\max\left(
+    \max(\kappa\mathbb{S}[Y(x_1)] - |y - Y(x_1)|,0),\dots,
+    \max(\kappa\mathbb{S}[Y(x_q)]
+     -|y - Y(x_q)|,0)
+    \right)\right]$$
+
+    where the expectation is taken with respect to the distribution of
+    the random vector $(Y(x_1),\dots,Y(x_q))$.
+    There is no analytic expression
+    and the acquisition is thus instead evaluated with crude Monte-Carlo.
     """
 
-    def _compute_output(self, input_value: NumberArray) -> NumberArray:  # noqa: D102
+    def _compute(self, input_value: NumberArray) -> NumberArray | float:  # noqa: D102
         # See Proposition 4, Bect et al, 2012
         standard_deviation, t, t_m, t_p = self._get_material(input_value)
         return nan_to_num(
@@ -64,3 +97,49 @@ class EF(BaseEIEF):
                 - (2 * norm.pdf(t) - norm.pdf(t_p) - norm.pdf(t_m))
             ),
         )
+
+    def _compute_empirically(  # noqa: D102
+        self, input_value: NumberArray
+    ) -> NumberArray | float:
+        self._regressor_distribution: RegressorDistribution
+        ndim_is_two = input_value.ndim == 2
+        input_data = atleast_2d(input_value)
+        samples = self._compute_samples(
+            input_data=input_data, n_samples=self._compute_samples
+        )
+        weights = self._regressor_distribution.evaluate_weights(input_data)
+        feasibility = self._kappa * self._compute_standard_deviation(input_value) - abs(
+            self._output_value - samples
+        )
+        expected_feasibility = maximum(feasibility, 0.0)
+        expected_feasibility = array([
+            dot(weights[:, index], expected_feasibility[:, index, :])
+            for index in range(expected_feasibility.shape[1])
+        ])
+        if ndim_is_two:
+            return nan_to_num(expected_feasibility)
+
+        return nan_to_num(expected_feasibility[0])
+
+    def _compute_by_batch(self, q_input_values: NumberArray) -> NumberArray | float:  # noqa: D102
+        q_input_values = self._reshape_input_values(q_input_values)
+        try:
+            samples = self._compute_samples(
+                input_data=q_input_values, n_samples=self._mc_size
+            )
+            return nan_to_num(
+                mean(
+                    np_max(
+                        maximum(
+                            self._kappa
+                            * self._compute_standard_deviation(q_input_values)
+                            - abs(self._output_value - samples),
+                            0,
+                        ),
+                        axis=1,
+                    ),
+                ),
+            )
+        # distribution is not positive definite.
+        except TypeError:
+            return 0.0