+++ src/gemseo_mlearning/active_learning/acquisition_criteria/maximum/ei.py
@@ -27,11 +27,37 @@ from gemseo_mlearning.active_learning.acquisition_criteria.maximum.base_maximum 
 class EI(ExpectedImprovement, BaseMaximum):
     r"""Expected improvement.
 
-    This acquisition criterion is expressed as:
+    This acquisition criterion is expressed as
 
     $$EI[x] = \mathbb{E}[\max(Y(x)-y_{\text{max}},0)]$$
 
     where $Y$ is the random process
     modelling the uncertainty of the surrogate model $\hat{f}$
     and $y_{\text{max}}$ is the maximum output values in the learning set.
+
+    In the case of a Gaussian regressor,
+    it has an analytic expression:
+
+    $$
+    EI[x] = (\mathbb{E}[Y(x)] - y_{\text{max}})\Phi(t)
+    + \mathbb{S}[Y(x)]\phi(t)
+    $$
+
+    where $Y$ is the random process
+    modelling the uncertainty of the surrogate model $\hat{f}$,
+    $t=\frac{\mathbb{E}[Y(x)] - y_{\text{max}}}{\mathbb{S}[Y(x)]}$
+    and $y_{\text{max}}$ is the maximum output value in the learning set.
+
+    For the acquisition of $q>1$ points at a time,
+    the acquisition criterion changes to
+
+    $$EI[x_1,\dots,x_q] = \mathbb{E}\left[\max\left(
+    \max(Y(x_1)-y_{\text{max}},0),\dots,
+    \max(Y(x_q)-y_{\text{max}},0)
+    \right)\right]$$
+
+    where the expectation is taken with respect to the distribution of
+    the random vector $(Y(x_1),\dots,Y(x_q))$.
+    There is no analytic expression
+    and the acquisition is thus instead evaluated with crude Monte-Carlo.
     """