+++ tests/active_learning/acquisition_criteria/test_maximum.py
@@ -18,6 +18,8 @@
 #    OTHER AUTHORS   - MACROSCOPIC CHANGES
 from __future__ import annotations
 
+import re
+
 import pytest
 from numpy import array
 from numpy.testing import assert_almost_equal
@@ -46,6 +48,8 @@ def test_maximum_kriging_regressor(
     """Check the criteria deriving from BaseMaximum with a Kriging distribution."""
     criterion = cls(kriging_distribution, **options)
     assert_almost_equal(criterion.func(input_value), expected)
+    assert criterion._mc_size == 10000
+    assert criterion._batch_size == 1
 
 
 @pytest.mark.parametrize(
@@ -65,3 +69,53 @@ def test_maximum_regressor(algo_distribution, options, cls, input_value, expecte
     """Check the criteria deriving from BaseMaximum with a RegressorDistribution."""
     criterion = cls(algo_distribution, **options)
     assert_almost_equal(criterion.func(input_value), expected)
+    assert criterion._mc_size == 10000
+    assert criterion._batch_size == 1
+
+
+@pytest.mark.parametrize(
+    ("options", "input_value", "expected"),
+    [
+        ({}, array([[1.0]] * 2), 0),
+        ({}, array([[3]] * 2), 0.0668520),
+        ({"mc_size": 50000}, array([[3]] * 2), 0.06647122),
+    ],
+)
+def test_maximum_parallel_kriging_regressor(
+    kriging_distribution, options, input_value, expected
+):
+    """Check the parallelized criteria deriving from BaseMaximum."""
+    criterion = EI(kriging_distribution, batch_size=2, **options)
+    assert_almost_equal(criterion.func(input_value), expected)
+    expected_mc_size = options.get("mc_size", 10000)
+    assert criterion._mc_size == expected_mc_size
+    assert criterion._batch_size == 2
+
+
+def test_improvement_parallel_at_training_point(kriging_distribution):
+    """Check that the improvement criteria take predefined values a training point."""
+    criterion = EI(kriging_distribution, 2)
+    criterion._compute_samples = lambda x: TypeError
+    assert_almost_equal(criterion.evaluate(array([0.0, 0.0])), 0)
+
+
+@pytest.mark.parametrize(
+    ("cls", "input_value"),
+    [
+        (EI, array([0.123])),
+        (UCB, array([0.123])),
+        (Output, array([0.123])),
+    ],
+)
+def test_bad_parallel_regressor(algo_distribution, cls, input_value):
+    """Check that parallelized criteria with non GP regressor lead to failure."""
+    criterion = cls(algo_distribution, batch_size=2)
+
+    with pytest.raises(
+        NotImplementedError,
+        match=re.escape(
+            "Parallelization with batch_size > 1 is not yet implemented "
+            "for regressors that are not based on a random process."
+        ),
+    ):
+        criterion.func(input_value)